Implement the initial logic for order payment

+++ ecommerce-parent/ecommerce-backend/src/main/java/com/iti/jets/repository/implementation/CartRepositoryImpl.java
@@ -188,7 +188,7 @@ public class CartRepositoryImpl extends BaseRepositoryImpl<Cart, Long> implement
     private Cart createCart(EntityManager em, User user) {
         Cart cart = Cart.builder()
                 .user(user)
-                .totalPrice(0)
+                .totalPrice((double) 0)
                 .build();
         em.persist(cart);
         em.flush();
